backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.14.1
anthropic==0.49.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
